Add main js and hit app running
