+++ .eslintrc.json
@@ -14,6 +14,8 @@
       "no-shadow": "off",
       "no-param-reassign": "off",
       "eol-last": "off",
+      "no-undef": "off",
+      "import/prefer-default-export": "off",
       "import/extensions": [ 1, {
         "js": "always", "json": "always"
       }]
